nuremberg/core/tests/requirements.txt
pyquery==1.2.13
pytest-django==2.9.1
pytest-cov==2.2.1
pytest-instafail-0.3.0
selenium==2.53.5
sure==1.2.24

-e git+https://github.com/emmalemma/django_coverage_plugin.git@e083da1624831c9485ec44f569d0689a59249f36#egg=django_coverage_plugin
